requirements.txt
httpx==0.24.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
